refactor: tests splitted
