+++ tests/__init__.py
@@ -1,1 +1,5 @@
 """Module for testing"""
+from module.shared import config_map
+
+TIMEOUT = 8
+bot_tag = config_map["test"]["tag"]
