+++ tests/test_start.py
@@ -1,0 +1,21 @@
+import pytest
+from telethon.sync import TelegramClient
+from telethon.tl.custom.conversation import Conversation
+from telethon.tl.custom.message import Message
+
+from . import TIMEOUT, bot_tag
+
+
+@pytest.mark.asyncio
+async def test_start_cmd(client: TelegramClient):
+    """Tests the /start command
+
+    Args:
+        client (TelegramClient): client used to simulate the user
+    """
+    conv: Conversation
+    async with client.conversation(bot_tag, timeout=TIMEOUT) as conv:
+        await conv.send_message("/start")  # send a command
+        resp: Message = await conv.get_response()
+
+        assert resp.text
